1021 knoct19_1 polsalt/saltobslog_kn.py        allow fits values mistakenly ints to be read as floats                                 remove sorting of infiles polsalt/specpolwavmap.py        implemented use of two simultaneous arcs 				allowed for odd number of rows polsalt/specpolsplit.py         allowed for odd number of rows polsalt/specpolrawstokes.py     added capability for multi-iteration exposures 				went from HWP-STA to HWP-ANG for wp posn 				remove wav0, wavs from config: allow for small xtract changes polsalt/data/spst_filenames.txt update to remove "_extr" from some names scripts/specpolflux.py          avoid error if no ec* fits present 				fix binning error scripts/specpolview.py          allow decimal Ang in binning

+++ scripts/specpolview.py
@@ -66,7 +66,7 @@ def specpolview(infile_list, **kwargs):
         errbin = float(bin[ :bin.index('%')])
     elif bin.count('A'): 
         bintype = 'wavl'
-        blk = int(bin[ :bin.index('A')])
+        blk = float(bin[ :bin.index('A')])
     elif (bin != 'unbin'): 
         print "unrecognized binning option, set to unbinned"
         bintype = 'unbin'
